Add automation for Bluesky content management via GitHub PRs
This adds several files for automating Bluesky content management.
The layout of files look like this:

```
- .github/workflows       # Contains workflows
    - process.yml         # Process merged requests
    - validate.yml        # Validate JSON requests added in PRs to catch early errors

- actions/       # scripts for the actions
    - lib/
    - test/
        - integration.js  # This can be run locally to test process.js on a test account
    - process.js  # a script that takes a path/to/new/name.json to perform actions
    - login-and-validate.js # part of process.js, can be run alone to validate a JSON

- records/
    - new/  # Where requests can be added in the form of JSON via pull requests.
            # It contains some existing examples with the name xxx.json.example
    - processed/  # Requests in the new folder will be moved here with information
                  # about the performed actions added. The file will be renamed
                  # to YYYY-MM-DD-ID.json, where ID is incremental index for files
                  # processed on that date.
```

To submit a request, send a PR to the main branch that adds new JSON
files to `records/new` (which already contain several `.json.example`
files as examples). The validation workflow is run on PRs adding new
JSON files to catch early errors. Once the PR is merged, the process
workflow will perform the requested action on configured accounts.

This PR also adds temporary review guidelines (only release
announcements are allowed at the moment) and documentation for the
automation.